CPM-1016: Refactor Acceptance tests
