+++ components/identifier-generator/back/tests/Acceptance/Context/BaseCreateOrUpdateIdentifierGenerator.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context;
+
+use Akeneo\Pim\Automation\IdentifierGenerator\Application\Create\CreateGeneratorCommand;
+use Akeneo\Pim\Automation\IdentifierGenerator\Application\Create\CreateGeneratorHandler;
+use Akeneo\Pim\Automation\IdentifierGenerator\Application\Exception\ViolationsException;
+use Akeneo\Pim\Automation\IdentifierGenerator\Application\Update\UpdateGeneratorCommand;
+use Akeneo\Pim\Automation\IdentifierGenerator\Application\Update\UpdateGeneratorHandler;
+
+/**
+ * @copyright 2023 Akeneo SAS (https://www.akeneo.com)
+ * @license   https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class BaseCreateOrUpdateIdentifierGenerator
+{
+    private const DEFAULT_IDENTIFIER_GENERATOR_CODE = 'default';
+
+    public function __construct(
+        private readonly ViolationsContext $violationsContext,
+        private readonly CreateGeneratorHandler $createGeneratorHandler,
+        private readonly UpdateGeneratorHandler $updateGeneratorHandler,
+    ) {
+    }
+
+    protected function tryToCreateGenerator(
+        ?string $code = null,
+        ?array $structure = null,
+        ?array $conditions = null,
+        ?array $labels = null,
+        ?string $target = null,
+        ?string $delimiter = null,
+        ?string $textTransformation = null,
+    ): void {
+        try {
+            ($this->createGeneratorHandler)(new CreateGeneratorCommand(
+                $code ?? self::DEFAULT_IDENTIFIER_GENERATOR_CODE,
+                $conditions ?? [],
+                $structure ?? [['type' => 'free_text', 'string' => self::DEFAULT_IDENTIFIER_GENERATOR_CODE]],
+                $labels ?? ['fr_FR' => 'Générateur'],
+                $target ?? 'sku',
+                $delimiter ?? '-',
+                $textTransformation ?? 'no',
+            ));
+        } catch (ViolationsException $exception) {
+            $this->violationsContext->setViolationsException($exception);
+        }
+    }
+
+    protected function tryToUpdateGenerator(
+        ?string $code = null,
+        ?array $structure = null,
+        ?array $conditions = null,
+        ?array $labels = null,
+        ?string $target = null,
+        ?string $delimiter = null,
+        ?string $textTransformation = null,
+    ): void {
+        try {
+            ($this->updateGeneratorHandler)(new UpdateGeneratorCommand(
+                $code ?? self::DEFAULT_IDENTIFIER_GENERATOR_CODE,
+                $conditions ?? [],
+                $structure ?? [['type' => 'free_text', 'string' => self::DEFAULT_IDENTIFIER_GENERATOR_CODE]],
+                $labels ?? ['fr_FR' => 'Générateur'],
+                $target ?? 'sku',
+                $delimiter ?? 'updatedGenerator',
+                $textTransformation ?? 'no',
+            ));
+        } catch (ViolationsException $violations) {
+            $this->violationsContext->setViolationsException($violations);
+        }
+    }
+}
