+++ components/identifier-generator/back/tests/Acceptance/Resources/config/services.yml
@@ -1,10 +1,39 @@
 services:
   # Context
+  Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\ViolationsContext:
+    public: true
+
+  Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\DatabaseContext:
+    public: true
+    arguments:
+      - '@Akeneo\Pim\Automation\IdentifierGenerator\Domain\Repository\IdentifierGeneratorRepository'
+
+  Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\Structure\AutoNumberContext:
+    public: true
+    arguments:
+      - '@Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\ViolationsContext'
+      - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Create\CreateGeneratorHandler'
+      - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Update\UpdateGeneratorHandler'
+
+  Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\Structure\FreeTextContext:
+    public: true
+    arguments:
+      - '@Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\ViolationsContext'
+      - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Create\CreateGeneratorHandler'
+      - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Update\UpdateGeneratorHandler'
+
+  Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\Structure\FamilyContext:
+    public: true
+    arguments:
+      - '@Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\ViolationsContext'
+      - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Create\CreateGeneratorHandler'
+      - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Update\UpdateGeneratorHandler'
+
   Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\CreateIdentifierGeneratorContext:
     public: true
     arguments:
+      - '@Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\ViolationsContext'
       - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Create\CreateGeneratorHandler'
-      - '@Akeneo\Pim\Automation\IdentifierGenerator\Domain\Repository\IdentifierGeneratorRepository'
       - '@pim_catalog.repository.attribute'
       - '@pim_catalog.repository.attribute_option'
       - '@Akeneo\Pim\Structure\Family\ServiceAPI\Query\FindFamilyCodes'
@@ -13,6 +42,7 @@ services:
   Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\UpdateIdentifierGeneratorContext:
     public: true
     arguments:
+      - '@Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\ViolationsContext'
       - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Update\UpdateGeneratorHandler'
       - '@Akeneo\Pim\Automation\IdentifierGenerator\Domain\Repository\IdentifierGeneratorRepository'
       - '@Akeneo\Pim\Automation\IdentifierGenerator\Application\Update\ReorderGeneratorsHandler'
