+++ components/identifier-generator/back/tests/behat.yml
@@ -4,6 +4,11 @@ default:
             paths:
                 - '%paths.base%/features'
             contexts:
+                - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\ViolationsContext
+                - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\DatabaseContext
+                - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\Structure\AutoNumberContext
+                - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\Structure\FreeTextContext
+                - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\Structure\FamilyContext
                 - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\CreateIdentifierGeneratorContext
                 - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\UpdateIdentifierGeneratorContext
                 - Akeneo\Test\Pim\Automation\IdentifierGenerator\Acceptance\Context\UpdateFamilyNomenclatureContext
